Adds interaction metadata for Activity Objects.
For Object#interactions, a hash is returned containing any known, and
thus normalized, information about that activity. If this is a
persisted object, one might have to be careful to update it. If it is
pulled from another server, it will generally contain information about
how it has interacted with other servers.

Likely the most obvious and useful statistic for a basic social system
is a count of the number of times something was shared or favorited:

object # Some Object
object.interactions # => {:share => {:count => 3,
                    #                :actors => ['id1', 'id2', 'id3']},
                    #     :like  => {:count => 1,
                    #                :actors => ['id4']}}
object.interactions[:share][:count] # => 3 (for instance)
object.interaction_count[:favorite] # => 0 (convenience function that
                                    #       imposes default of 0)