Removes ratom dependency by parsing Atom based AS1 ourselves.
I was hacking around the library too much and it was just... breaking
horribly. This replacement is basically as much if not less code to do
the same thing and it supports a wider range of AS1 Atom/XML
representations.

For instance, it does NOT break on status.net's more AS1 happy feed
syndication. *confetti* It parses it all correctly. Woo.